mister-companion/requirements.txt
PyQt6
paramiko
requests
websocket-client
psutil